lms-requirements.txt
django-model-utils==4.4.0
django-crispy-forms==2.1
Pillow==10.2.0
python-dotenv==1.0.1
markdown==3.5.2
